feat: scratchpads & ui improvements

+++ source/editor/default_editor.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <cstdlib>
+#include <filesystem>
 #include <fstream>
 #include <random>
 #include <string>
@@ -15,40 +16,68 @@ namespace caps_log::editor {
 
 class DefaultEditor : public EditorBase {
     log::LocalFSLogFilePathProvider m_pathProvider;
+    std::filesystem::path m_scratchpadDirPath;
     std::string m_editorCommand;
 
   public:
     explicit DefaultEditor(caps_log::log::LocalFSLogFilePathProvider pathProvider,
-                           std::string editorCommand)
-        : m_pathProvider{std::move(pathProvider)}, m_editorCommand{std::move(editorCommand)} {}
-    void openEditor(const caps_log::log::LogFile &log) override {
-        if (std::getenv("EDITOR") != nullptr) {
-            std::ignore =
-                std::system(("$EDITOR " + m_pathProvider.path(log.getDate()).string()).c_str());
-        }
+                           std::filesystem::path scratchpadDirPath, std::string editorCommand)
+        : m_pathProvider{std::move(pathProvider)},
+          m_scratchpadDirPath{std::move(scratchpadDirPath)},
+          m_editorCommand{std::move(editorCommand)} {}
+
+    void openLog(const caps_log::log::LogFile &log) override {
+        auto editorOpenCommand =
+            fmt::format("{} \"{}\"", m_editorCommand, m_pathProvider.path(log.getDate()).string());
+        std::ignore = std::system(editorOpenCommand.c_str());
+    }
+
+    void openScratchpad(const std::string &scratchpadName) override {
+        auto editorOpenCommand = fmt::format("{} \"{}\"", m_editorCommand,
+                                             (m_scratchpadDirPath / scratchpadName).string());
+        std::ignore = std::system(editorOpenCommand.c_str());
     }
 };
 
 class EncryptedDefaultEditor : public EditorBase {
     caps_log::log::LocalFSLogFilePathProvider m_pathProvider;
+    std::filesystem::path m_scratchpadPath;
     std::string m_password;
     std::string m_editorCommand;
 
   public:
     EncryptedDefaultEditor(caps_log::log::LocalFSLogFilePathProvider pathProvider,
-                           std::string password, std::string editorCommand)
-        : m_pathProvider{std::move(pathProvider)}, m_password{std::move(password)},
-          m_editorCommand{std::move(editorCommand)} {}
+                           std::filesystem::path scratchpadPath, std::string password,
+                           std::string editorCommand)
+        : m_pathProvider{std::move(pathProvider)}, m_scratchpadPath{std::move(scratchpadPath)},
+          m_password{std::move(password)}, m_editorCommand{std::move(editorCommand)} {}
 
-    void openEditor(const caps_log::log::LogFile &log) override {
+    void openLog(const caps_log::log::LogFile &log) override {
         const auto tmp = getTmpFile();
         const auto originalLogPath = m_pathProvider.path(log.getDate());
-        std::filesystem::copy_file(originalLogPath, tmp,
-                                   std::filesystem::copy_options::overwrite_existing);
-        decryptFile(tmp);
-        openEditor(tmp);
+        if (std::filesystem::exists(originalLogPath)) {
+            std::filesystem::copy_file(originalLogPath, tmp,
+                                       std::filesystem::copy_options::overwrite_existing);
+            decryptFile(tmp);
+        }
+        openRawEditor(tmp);
         encryptFile(tmp);
         std::filesystem::copy_file(tmp, originalLogPath,
+                                   std::filesystem::copy_options::overwrite_existing);
+    }
+
+    void openScratchpad(const std::string &scratchpadName) override {
+        const auto tmp = getTmpFile();
+        const auto originalScratchpadPath = m_scratchpadPath / scratchpadName;
+        if (std::filesystem::exists(originalScratchpadPath)) {
+            // Copy the original scratchpad to a temporary file
+            std::filesystem::copy_file(originalScratchpadPath, tmp,
+                                       std::filesystem::copy_options::overwrite_existing);
+            decryptFile(tmp);
+        }
+        openRawEditor(tmp);
+        encryptFile(tmp);
+        std::filesystem::copy_file(tmp, originalScratchpadPath,
                                    std::filesystem::copy_options::overwrite_existing);
     }
 
@@ -89,8 +118,9 @@ class EncryptedDefaultEditor : public EditorBase {
         }
     }
 
-    void openEditor(const std::string &path) {
-        std::ignore = std::system((m_editorCommand + " " + path).c_str());
+    void openRawEditor(const std::string &path) {
+        auto editorOpenCommand = fmt::format("{} \"{}\"", m_editorCommand, path);
+        std::ignore = std::system(editorOpenCommand.c_str());
     }
 };
